third_party/blink/web_tests/accessibility/element-role-mapping-normal.html

<html>
<meta charset="utf-8">
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
<body>
<div id="container">
<header>This is a Header for this page</header>
<h1>Heading</h1>
<div>Division</div>
<p>Paragraph</p>
<ruby>
韓國<rp>(</rp><rt>한국</rt><rp>)</rp>
</ruby>
<dl>
  <dt>Coffee</dt>
  <dd>- black hot drink</dd>
  <dt>Milk</dt>
  <dd>- white cold drink</dd>
</dl>
<math role="math" xmlns="http://www.w3.org/1998/Math/MathML">
  <mrow>
    <mi> x </mi>
    <mo> + </mo>
    <mrow>
      <mi> a </mi>
      <mo> / </mo>
      <mi> b </mi>
    </mrow>
  </mrow>
</math>
<math xmlns="http://www.w3.org/1998/Math/MathML">
  <mrow>
    <mi>x</mi>
    <mo>+</mo>
    <mrow>
      <mi>a</mi>
      <mo>/</mo>
      <mi>b</mi>
    </mrow>
  </mrow>
</math>
<main>
  <article>
    <header>
      <h1>Most important heading here</h1>
    </header>
    <h1>Google Chrome</h1>
    <p>Google Chrome is a free, open-source web browser developed by Google,
    released in 2008.</p>
    <footer>
      <p>Footer in article</p>
    </footer>
  </article>
</main>
<nav>
  <a href="/html/">HTML</a> |
  <a href="/css/">CSS</a> |
  <a href="/js/">JavaScript</a> |
</nav>
<aside>
  <p>Text that appears under aside</p>
</aside>
<pre>
Text in a    pre
element
</pre>
<section>
  <header>
    <h1>Most important heading here</h1>
  </header>
  <p>Section</p>
  <footer>
    <p>Footer in section</p>
  </footer>
</section>
<address>
Written by Julie<br>
Visit us at:www.chromium.org<br>
</address>
<hgroup>
  <h1>Main Heading</h1>
  <h5>Subheading</h5>
  <p>Smaller Subheading</p>
</hgroup>
<p>January <dialog open>This is an open dialog window</dialog></p>
<table>
  <caption>Caption</caption>
  <tr>
    <td>Cell1</td>
    <td>Cell2</td>
  </tr>
</table>
<figure>
  <img width="12" height="10" src="data:image/gif;base64,R0lGODdhDAAKAJEAAP///wAA/wAAAAAAACwAAAAADAAKAAACEoyPAMihnRY0bzo5a8ZQd94oBQA7" alt="blue">
  <figcaption>Fig1. - Blue Box</figcaption>
</figure>
<embed src="xxx.mp3">
<footer>
  <p>This is a footer.</p>
</footer>
</div>  <!-- container -->
<p id="description"></p>
<div id="console"></div>
<script>
description("This test make sure that unfocuasable elements are mapped implicitly to AX roles.");

if (window.testRunner)
  testRunner.dumpAsText();
if (window.accessibilityController)
    buildAccessibilityTree(accessibilityController.accessibleElementById("container"), 0, 1);
</script>
</body>
</html>
